+++ src/image-text/image-text.js
@@ -4,7 +4,11 @@ import DOMPurify from "dompurify";
 import PropTypes from "prop-types";
 import { CSSTransition, TransitionGroup } from "react-transition-group";
 import BaseSlideExecution from "../base-slide-execution";
-import { getAllMediaUrlsFromField, ThemeStyles } from "../slide-util";
+import {
+  getAllMediaUrlsFromField,
+  getFirstMediaUrlFromField,
+  ThemeStyles,
+} from "../slide-util";
 import "../global-styles.css";
 import "./image-text.scss";
 
@@ -24,6 +28,27 @@ function ImageText({ slide, content, run, slideDone, executionId }) {
   const [images, setImages] = useState([]);
   const [currentImage, setCurrentImage] = useState();
   const [themeCss, setThemeCss] = useState(null);
+  const logo = slide?.themeData?.logo;
+  const { showLogo, logoSize, logoPosition, logoMargin } = content;
+  const { disableImageFade } = content;
+
+  let logoUrl = "";
+  // If showlogo is set, get the logo url
+  if (logo && showLogo) {
+    logoUrl = getFirstMediaUrlFromField(slide.mediaData, [logo]);
+  }
+
+  const logoClasses = ["logo"];
+
+  if (logoMargin) {
+    logoClasses.push("logo-margin");
+  }
+  if (logoSize) {
+    logoClasses.push(logoSize);
+  }
+  if (logoPosition) {
+    logoClasses.push(logoPosition);
+  }
 
   // Set theme styles.
   useEffect(() => {
@@ -147,7 +172,7 @@ function ImageText({ slide, content, run, slideDone, executionId }) {
       setCurrentImage(images[0]);
     }
 
-    // If there are multiple images, we are going to loop through these WITHIN the set duration. 
+    // If there are multiple images, we are going to loop through these WITHIN the set duration.
     if (images?.length > 1) {
       // Kickoff the display of multiple images with the zero indexed
       changeImage(0);
@@ -180,7 +205,9 @@ function ImageText({ slide, content, run, slideDone, executionId }) {
               key={currentImage.url}
               timeout={1000}
               nodeRef={currentImage.nodeRef}
-              classNames="background-image"
+              classNames={`background-image${
+                disableImageFade ? "-animation-disabled" : ""
+              }`}
             >
               <div
                 style={{
@@ -208,7 +235,12 @@ function ImageText({ slide, content, run, slideDone, executionId }) {
             )}
           </div>
         )}
+
+        {showLogo && logoUrl && (
+          <img className={logoClasses.join(" ")} src={logoUrl} alt="" />
+        )}
       </div>
+
       {themeCss}
     </>
   );
@@ -224,6 +256,7 @@ ImageText.propTypes = {
     }),
     themeData: PropTypes.shape({
       cssStyles: PropTypes.string,
+      logo: PropTypes.string,
     }),
   }).isRequired,
   content: PropTypes.shape({
@@ -243,6 +276,11 @@ ImageText.propTypes = {
       halfSize: PropTypes.bool,
       fontSize: PropTypes.string,
     }),
+    showLogo: PropTypes.bool,
+    logoSize: PropTypes.string,
+    logoMargin: PropTypes.bool,
+    logoPosition: PropTypes.string,
+    disableImageFade: PropTypes.bool,
   }).isRequired,
   executionId: PropTypes.string.isRequired,
 };