+++ src/slideshow/slideshow.js
@@ -1,6 +1,10 @@
 import React, { useState, useRef, useEffect } from "react";
 import PropTypes from "prop-types";
-import { getAllMediaUrlsFromField, ThemeStyles } from "../slide-util";
+import {
+  getAllMediaUrlsFromField,
+  getFirstMediaUrlFromField,
+  ThemeStyles,
+} from "../slide-util";
 import "../global-styles.css";
 import "./slideshow.scss";
 
@@ -16,34 +20,44 @@ import "./slideshow.scss";
  * @returns {JSX.Element} The component.
  */
 function Slideshow({ slide, content, run, slideDone, executionId }) {
-  const {
-    images,
-    imageDuration = 5000,
-    transitions,
-    animations,
-    // @TODO: Add when logo is available from theme
-    // , logoEnabled, logoSize, logoPosition
-  } = content;
+  const { images, imageDuration = 5000, transition, animation } = content;
   const [index, setIndex] = useState(0);
-  const fadeEnabled = transitions === "fade";
+  const fadeEnabled = transition === "fade";
   const fadeDuration = 1000;
   const [fade, setFade] = useState(false);
-
+  const imageDurationInSeconds = imageDuration * 1000;
   // Map images to mediaData.
   const imageUrls = getAllMediaUrlsFromField(slide.mediaData, images);
 
   const animationName = "animationForImage";
   const [animationIndex, setAnimationIndex] = useState(0);
   const [animationDuration, setAnimationDuration] = useState(
-    imageDuration + fadeDuration
+    imageDurationInSeconds + fadeDuration
   );
+
+  const logo = slide?.themeData?.logo;
+  const { showLogo, logoSize, logoPosition, logoMargin } = content;
+
+  let logoUrl = "";
+  // If showlogo is set, get the logo url
+  if (logo && showLogo) {
+    logoUrl = getFirstMediaUrlFromField(slide.mediaData, [logo]);
+  }
+
+  const logoClasses = ["logo"];
+
+  if (logoMargin) {
+    logoClasses.push("logo-margin");
+  }
+  if (logoSize) {
+    logoClasses.push(logoSize);
+  }
+  if (logoPosition) {
+    logoClasses.push(logoPosition);
+  }
 
   const timeoutRef = useRef(null);
   const fadeRef = useRef(null);
-
-  // @TODO: Get logo from theme.
-  // const logoImageUrl = null;
-  // const logoClasses = `logo ${logoPosition} ${logoSize}`;
 
   /**
    * A random function to simplify the code where random is used
@@ -102,7 +116,6 @@ function Slideshow({ slide, content, run, slideDone, executionId }) {
     ];
 
     const randomPercent = `${random(100) + 1}% ${random(100) + 1}%`;
-
     switch (animationType) {
       case "zoom-in-middle":
         return createAnimation(true);
@@ -112,54 +125,58 @@ function Slideshow({ slide, content, run, slideDone, executionId }) {
         return createAnimation(true, randomPercent);
       case "zoom-out-random":
         return createAnimation(false, randomPercent);
-      default:
       case "random":
         return getCurrentAnimation(
           animationTypes[random(animationTypes.length)]
         );
+      default:
+        return null;
     }
   }
 
   // Setup animation
   useEffect(() => {
-    if (animations !== null) {
+    if (animation !== null) {
       // Adds the animation to the stylesheet. because there is an element of random, we cannot have it in the .scss file.
       const styleSheet = document.styleSheets[0];
-      styleSheet.insertRule(
-        getCurrentAnimation(animations),
-        styleSheet.cssRules.length
-      );
+      const currentAnimation = getCurrentAnimation(animation);
+      if (currentAnimation !== null) {
+        styleSheet.insertRule(
+          getCurrentAnimation(animation),
+          styleSheet.cssRules.length
+        );
+      }
     }
   }, []);
 
   // Get image style for the given image url.
-  const getImageStyle = (imageUrl, animation, localAnimationDuration) => {
+  const getImageStyle = (imageUrl, localAnimation, localAnimationDuration) => {
     const imageStyle = {
       backgroundImage: `url(${imageUrl})`,
     };
 
-    if (animation) {
+    if (localAnimation) {
       imageStyle.animation = `${animationName} ${localAnimationDuration}ms`;
     }
 
     return imageStyle;
   };
-
   // Setup image progress.
   useEffect(() => {
     if (run) {
       if (imageUrls.length > 0) {
         timeoutRef.current = setTimeout(() => {
-          const newIndex = index + 1;
-
-          if (newIndex > imageUrls.length - 1) {
+          let newIndex = index + 1;
+          if (newIndex === imageUrls.length) {
+            newIndex = 0;
             // No more images to show.
             slideDone(slide);
-          } else if (fadeEnabled) {
+          }
+          if (fadeEnabled) {
             // Fade to next image.
             setFade(true);
             setAnimationIndex(newIndex);
-            setAnimationDuration(imageDuration + fadeDuration * 2);
+            setAnimationDuration(imageDurationInSeconds + fadeDuration * 2);
             fadeRef.current = setTimeout(() => {
               setFade(false);
               setIndex(newIndex);
@@ -168,9 +185,9 @@ function Slideshow({ slide, content, run, slideDone, executionId }) {
             // Change to next.
             setIndex(newIndex);
             setAnimationIndex(newIndex);
-            setAnimationDuration(imageDuration);
+            setAnimationDuration(imageDurationInSeconds);
           }
-        }, imageDuration);
+        }, imageDurationInSeconds);
       } else {
         // If there are no images in slide, wait for 2s before continuing to avoid crashes.
         setTimeout(() => {
@@ -230,6 +247,10 @@ function Slideshow({ slide, content, run, slideDone, executionId }) {
               </div>
             );
           })}
+
+        {showLogo && logoUrl && (
+          <img className={logoClasses.join(" ")} src={logoUrl} alt="" />
+        )}
       </div>
 
       <ThemeStyles id={executionId} css={slide?.themeData?.cssStyles} />
@@ -246,17 +267,19 @@ Slideshow.propTypes = {
       assets: PropTypes.shape({ uri: PropTypes.string }),
     }),
     themeData: PropTypes.shape({
-      css: PropTypes.string,
+      cssStyles: PropTypes.string,
+      logo: PropTypes.string,
     }),
   }).isRequired,
   content: PropTypes.shape({
     images: PropTypes.arrayOf(PropTypes.string),
     imageDuration: PropTypes.number,
-    logoEnabled: PropTypes.bool,
+    animation: PropTypes.string,
+    transition: PropTypes.string,
+    showLogo: PropTypes.bool,
     logoSize: PropTypes.string,
+    logoMargin: PropTypes.bool,
     logoPosition: PropTypes.string,
-    animations: PropTypes.string,
-    transitions: PropTypes.string,
   }).isRequired,
   executionId: PropTypes.string.isRequired,
 };